\name{RVineMatrix}
\alias{RVineMatrix}
\alias{add.VineMatrix}
\title{Notation of the R-vine Matrix}
\description{Notation of the R-vine Matrix.}
\usage{RVineMatrix(help.env)}
\arguments{
  \item{help.env}{Environment containing all information, environment of
    vine()}
}

\value{RVineMatrix}

\references{
  Nonparametric estimation of simplified vines: comparison of methods, Nagler N., Schellhase, C. and Czado, C. (2017) Dependence Modeling.
 }
\author{Christian Schellhase <cschellhase@wiwi.uni-bielefeld.de>}
